+++ NEWS.md
@@ -25,3 +25,8 @@
 * Added function defCondition - define distribution conditional on existing fields
 * Added function defReadCond - read in conditional definitions from external csv file
 * Added function addCondition - genaration data based on conditional definition
+
+# simstudy 0.1.5
+
+* Added integer uniform distribution (uniformInt)
+* Added negative binomial distribution (negBinomial)